Refactor project structure packaging
Packages are now roughly grouped by feature as suggested by the
developer documentation. This should hopefully add ease of navigation
and maintenance.
