+++ app/src/main/AndroidManifest.xml
@@ -9,19 +9,19 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".activities.MainActivity">
+        <activity android:name=".main.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".activities.AddTask" />
+        <activity android:name=".taskutil.AddTaskActivity" />
         <activity
-            android:name=".activities.FlowOverview"
+            android:name=".flow.FlowOverviewActivity"
             android:theme="@style/Theme.AppCompat.NoActionBar" />
-        <activity android:name=".activities.FlowSettings" />
-        <activity android:name=".activities.EditTask" />
+        <activity android:name=".main.SettingsActivity" />
+        <activity android:name=".taskutil.EditTaskActivity" />
     </application>
 
 </manifest>
