+++ app/src/main/java/com/sudosays/tempo/taskutil/TaskView.kt
@@ -1,4 +1,4 @@
-package com.sudosays.tempo.views
+package com.sudosays.tempo.taskutil
 
 /**
  *
