+++ app/src/main/res/layout/activity_flow_overview.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.sudosays.tempo.activities.FlowOverview"
+    tools:context="com.sudosays.tempo.flow.FlowOverviewActivity"
     android:id="@+id/flowParent">
 
     <TextView
