+++ app/src/main/res/layout/activity_flow_settings.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.sudosays.tempo.activities.FlowSettings">
+    tools:context="com.sudosays.tempo.main.SettingsActivity">
 
     <SeekBar
         android:id="@+id/taskTimeSeekBar"
